유저 휴면, 삭제처리

+++ src/main/java/com/handong/cra/crawebbackend/auth/service/AuthServiceImpl.java
@@ -8,6 +8,7 @@ import com.handong.cra.crawebbackend.auth.dto.SignupDto;
 import com.handong.cra.crawebbackend.auth.dto.TokenDto;
 import com.handong.cra.crawebbackend.auth.dto.response.ResTokenDto;
 import com.handong.cra.crawebbackend.auth.repository.RefreshTokenRepository;
+import com.handong.cra.crawebbackend.exception.user.UserDormantUserLoginException;
 import com.handong.cra.crawebbackend.user.dto.LoginUserDto;
 import com.handong.cra.crawebbackend.user.dto.UserDetailDto;
 import com.handong.cra.crawebbackend.user.repository.UserRepository;
@@ -77,6 +78,12 @@ public class AuthServiceImpl implements AuthService {
 
         // 로그인시 전달할 유저의 정보
         UserDetailDto userDetailDto = userService.getUserDetailByUsername(loginUserDto.getUsername());
+
+        // 휴면 계정 로그인시
+        if (userDetailDto.getDeleted()) throw new UserDormantUserLoginException();
+
+        //로그인 시간 갱신
+        userService.setLoginTimeById(userDetailDto.getId());
 
         // 이미 존재하면 삭제
         refreshTokenRepository.deleteAllByUserId(userDetailDto.getId());
